game works! let's add another DOF

+++ src/main.rs
@@ -18,8 +18,8 @@ const TIOCGWINSZ: c_int = 21523;
 const WIDTH: usize = 50;
 const HEIGHT: usize = 30;
 // difficulty attributes (inversely proportional to difficulty)
-const TIMEOUT_MS: c_uint = 100;     // gameplay speed
-const CLIFF_SEPARATION: u8 = 6;     // throw cliffs every X lines
+const TIMEOUT_MS: c_uint = 100;         // gameplay speed
+const CLIFF_SEPARATION: usize = 6;      // throw cliffs every X lines
 
 fn main() {
     let _raw = match set_raw_mode() {   // old termios attributes (which will be restored on drop)
